requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
django-filter>=24.0,<25.0
django-environ>=0.11,<0.12
channels[daphne]>=4.1,<4.2
channels-redis>=4.2,<4.3
psycopg2-binary>=2.9,<2.10
Pillow>=10.2,<10.3
celery>=5.3,<5.4
django-celery-beat>=2.6,<2.7
redis>=5.0,<5.1
gunicorn>=21.2,<21.3
drf-spectacular>=0.27,<0.28
django-storages>=1.14,<1.15
boto3>=1.34,<1.35
requests>=2.31,<2.32
geopy>=2.4,<2.5
firebase-admin>=6.4,<6.5
twilio>=9.0,<9.1
whitenoise>=6.6,<6.7
django-import-export>=3.3,<3.4
